risk_management_service/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.5.1
requests==2.31.0
cryptography==41.0.7
python-dotenv==1.0.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
numpy==1.24.3
pandas==2.0.3
psutil==5.9.6
